refactor(#10):: Simplify response result creation with new content constructors
